requirements.txt
annotated-types==0.6.0
anyio~=3.6.2
APScheduler==3.6.3
certifi~=2022.12.7
cffi==1.16.0
cryptography==41.0.7
decorator==5.1.1
distro==1.8.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna~=3.4
openai==1.3.6
pycparser==2.21
pydantic~=2.4.2
pydantic_core==2.14.5
python-telegram-bot==13.0
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.8.0
tzlocal==5.2
pip~=23.3.1
attrs~=22.2.0
wheel~=0.38.4
setuptools~=65.5.1
requests~=2.28.2
rich~=13.3.3
Pygments~=2.14.0